+++ app/seeds/reward_items.py
@@ -62,8 +62,25 @@ def seed_reward_items():
 # incrementing primary key, CASCADE deletes any dependent entities.  With
 # sqlite3 in development you need to instead use DELETE to remove all data and
 # it will reset the primary keys for you as well.
+
+def does_table_exist(table_name, schema_name):
+    query = text(
+        """
+        SELECT EXISTS (
+            SELECT 1
+            FROM information_schema.tables
+            WHERE table_schema = :schema AND table_name = :table
+        );
+        """
+    )
+
+    result = db.session.execute(query, {'schema': schema_name, 'table': table_name})
+    exists = result.scalar()
+    return exists
+
 def undo_reward_items():
     if environment == "production":
+      if does_table_exist("reward_items", SCHEMA):
         db.session.execute(f"TRUNCATE table {SCHEMA}.reward_items RESTART IDENTITY CASCADE;")
     else:
         db.session.execute(text("DELETE FROM reward_items"))
